refactor: define expiresAt as int64

+++ lexfloatclient.go
@@ -27,7 +27,7 @@ type HostFeatureEntitlement struct {
 	FeatureName          string `json:"featureName"`
 	FeatureDisplayName   string `json:"featureDisplayName"`
 	Value                string `json:"value"`
-	ExpiresAt            uint   `json:"expiresAt"`
+	ExpiresAt            int64  `json:"expiresAt"`
 }
 
 type callbackType func(int)
